Add ServiceOrchestrationMixin
